requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
gunicorn==20.0.4
static3==0.7.0
psycopg2 >= 2.8.5
django-ckeditor ==5.9.0
pillow == 7.1.2
